Extract components with queries
