+++ src/pages/index.js
@@ -1,20 +1,15 @@
 import React from 'react'
-import { StaticQuery, graphql } from 'gatsby'
-import get from 'lodash.get'
 
 import Layout from '../components/Layout'
-import Carousel from '../components/Carousel'
 import ConstrainedWidth from '../components/Layout/ConstrainedWidth'
 import Text from '../components/Text'
 import VerticalSpacing from '../components/VerticalSpacing'
 import Heading from '../components/Heading'
 import Link, { ExternalLink } from '../components/Link'
-import { WorkSummary } from '../components/WorkDescription'
 
 import SquiggleSeparator from '../components/SquiggleSeparator'
-import RenderContent from '../components/RenderContent'
-
-import logo from '../components/Logo/logo.svg'
+import StoryCarousel from '../components/StoryCarousel'
+import HomePageWorkSummary from '../components/HomePageWorkSummary'
 
 const randomItem = items => items[Math.floor(Math.random() * items.length)]
 
@@ -43,78 +38,7 @@ const IndexPage = () => (
             The tribe
           </Text>
         </Heading>
-        <StaticQuery
-          query={graphql`
-            query {
-              triberStories: allMarkdownRemark(
-                sort: { fields: [fileAbsolutePath] }
-                filter: {
-                  fields: { sourceName: { eq: "tribers" } }
-                  frontmatter: {
-                    story: { id: { ne: null } }
-                    type: { ne: "story" }
-                  }
-                }
-              ) {
-                nodes {
-                  frontmatter {
-                    image {
-                      childImageSharp {
-                        fluid(
-                          maxWidth: 400
-                          maxHeight: 420
-                          cropFocus: CENTER
-                        ) {
-                          src
-                        }
-                      }
-                    }
-                    name
-                    story {
-                      md: childMarkdownRemark {
-                        frontmatter {
-                          name
-                        }
-                        htmlAst
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          `}
-          render={data => {
-            const tribers = data.triberStories.nodes.slice(0, 3)
-
-            return (
-              <Carousel>
-                {tribers.map(triber => (
-                  <Carousel.Item
-                    key={triber.frontmatter.name}
-                    image={
-                      get(
-                        triber,
-                        'frontmatter.image.childImageSharp.fluid.src'
-                      ) || logo
-                    }
-                  >
-                    <RenderContent
-                      htmlAst={triber.frontmatter.story.md.htmlAst}
-                      type="secondary"
-                    />
-                    <VerticalSpacing size={1} />
-                    <Heading level={3} size="medium">
-                      <Text size="medium" type="secondary" heavy>
-                        {triber.frontmatter.story.md.frontmatter.name ||
-                          triber.frontmatter.name}
-                      </Text>
-                    </Heading>
-                  </Carousel.Item>
-                ))}
-              </Carousel>
-            )
-          }}
-        />
+        <StoryCarousel />
       </section>
 
       <section className="centered">
@@ -133,46 +57,8 @@ const IndexPage = () => (
             </Text>
           </Heading>
         </div>
-
         <VerticalSpacing size={5} />
-
-        <StaticQuery
-          query={graphql`
-            query {
-              projects: allMarkdownRemark(
-                filter: { fields: { sourceName: { eq: "projects" } } }
-              ) {
-                nodes {
-                  frontmatter {
-                    title
-                    problem
-                    image {
-                      childImageSharp {
-                        fluid(maxWidth: 768) {
-                          ...GatsbyImageSharpFluid
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          `}
-          render={data => {
-            const randProj = randomItem(data.projects.nodes)
-
-            return (
-              <WorkSummary
-                level={2}
-                title={randProj.frontmatter.title}
-                problem={randProj.frontmatter.problem}
-                fluid={randProj.frontmatter.image.childImageSharp.fluid}
-                alternate
-              />
-            )
-          }}
-        />
-
+        <HomePageWorkSummary />
         <VerticalSpacing size={5} />
 
         <div className="centered">
